Refactor: move slack_to_human_readable to new module
