+++ response/core/serializers.py
@@ -1,7 +1,7 @@
 from rest_framework import serializers
 
 from response.core.models import Action, ExternalUser, Incident, TimelineEvent
-from response.slack.client import slack_to_human_readable
+from response.slack.reference_utils import slack_to_human_readable
 from response.slack.models import CommsChannel
